requirements.txt
openmeteo_requests
openmeteo_sdk
requests
requests-cache
retry-requests
setuptools
pyyaml
pandas